+++ backend/canisters/local_user_index/impl/src/model/global_user_map.rs
@@ -63,6 +63,20 @@ impl GlobalUserMap {
         }
     }
 
+    pub fn remove(&mut self, user_id: &UserId) -> bool {
+        if let Some(principal) = self.user_id_to_principal.remove(user_id) {
+            if self.principal_to_user_id.get(&principal) == Some(user_id) {
+                self.principal_to_user_id.remove(&principal);
+            }
+            self.platform_moderators.remove(user_id);
+            self.bots.remove(user_id);
+            self.diamond_membership_expiry_dates.remove(user_id);
+            true
+        } else {
+            false
+        }
+    }
+
     pub fn is_bot(&self, user_id: &UserId) -> bool {
         self.bots.contains(user_id)
     }