+++ frontend/openchat-shared/src/domain/worker.ts
@@ -162,10 +162,7 @@ import type {
     SiwePrepareLoginResponse,
     SiwsPrepareLoginResponse,
 } from "./identity";
-import type {
-    GenerateEmailVerificationCodeResponse,
-    SubmitEmailVerificationCodeResponse,
-} from "./email";
+import type { GenerateMagicLinkResponse } from "./email";
 import type {
     ApproveResponse,
     MarkDeployedResponse,
@@ -354,8 +351,7 @@ export type WorkerRequest =
     | GetLocalUserIndexForUser
     | UpdateBtcBalance
     | SetPrincipalMigrationJobEnabled
-    | GenerateEmailVerificationsCode
-    | SubmitEmailVerificationsCode
+    | GenerateMagicLink
     | GetSignInWithEmailDelegation
     | SiwePrepareLogin
     | SiwsPrepareLogin
@@ -1172,16 +1168,10 @@ type SetPrincipalMigrationJobEnabled = {
     kind: "setPrincipalMigrationJobEnabled";
 };
 
-type GenerateEmailVerificationsCode = {
+type GenerateMagicLink = {
     email: string;
-    kind: "generateEmailVerificationCode";
-};
-
-type SubmitEmailVerificationsCode = {
-    email: string;
-    code: string;
     sessionKey: Uint8Array;
-    kind: "submitEmailVerificationCode";
+    kind: "generateMagicLink";
 };
 
 type GetSignInWithEmailDelegation = {
@@ -1359,8 +1349,7 @@ export type WorkerResponseInner =
     | PendingDeploymentResponse
     | JoinVideoCallResponse
     | UpdateBtcBalanceResponse
-    | GenerateEmailVerificationCodeResponse
-    | SubmitEmailVerificationCodeResponse
+    | GenerateMagicLinkResponse
     | SiwePrepareLoginResponse
     | SiwsPrepareLoginResponse
     | SetVideoCallPresenceResponse
@@ -1992,10 +1981,8 @@ export type WorkerResult<T> = T extends PinMessage
     ? UpdateBtcBalanceResponse
     : T extends SetPrincipalMigrationJobEnabled
     ? void
-    : T extends GenerateEmailVerificationsCode
-    ? GenerateEmailVerificationCodeResponse
-    : T extends SubmitEmailVerificationsCode
-    ? SubmitEmailVerificationCodeResponse
+    : T extends GenerateMagicLink
+    ? GenerateMagicLinkResponse
     : T extends GetSignInWithEmailDelegation
     ? GetDelegationResponse
     : T extends SiwePrepareLogin